+++ src/Console/Commands/ChannelMakeCommand.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace LaravelModular\LaravelModular\Console\Commands;
+namespace Laltu\Modular\Console\Commands;
 
 use Illuminate\Foundation\Console\ChannelMakeCommand as BaseChannelMakeCommand;