+++ src/Console/Commands/ModelMakeCommand.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace LaravelModular\LaravelModular\Console\Commands;
+namespace Laltu\Modular\Console\Commands;
 
 use Illuminate\Foundation\Console\ModelMakeCommand as BaseModelMakeCommand;
 use Illuminate\Support\Str;
-use LaravelModular\LaravelModular\Support\Config;
+use Laltu\Modular\Support\Config;
 
 final class ModelMakeCommand extends BaseModelMakeCommand
 {